Refactor each command into its own file. Created utils for common functions.
